+++ src/pages/ProductPage/index.jsx
@@ -2,6 +2,10 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
+import ProductQuantity from 'components/ProductQuantity';
+
+import Skeleton from 'skeletons/ProductDetailSkeleton';
+
 import Wrapper from './style';
 
 import { getProduct } from 'reducers/product';
@@ -13,21 +17,21 @@ const ProductPage = () => {
   const { loading: cartLoading, data: cart } = useSelector((state) => state.cart);
   const { productId } = useParams();
 
-  console.log(product, cart, productLoading, cartLoading);
-
   useEffect(() => {
     dispatch(getProduct(productId));
     dispatch(getCart(productId));
   }, [dispatch, productId]);
 
-  if (productLoading || cartLoading) return '로딩';
+  if (productLoading || cartLoading) return <Skeleton />;
 
   return (
     <Wrapper>
       <div className="product-wrapper">
         <img src={product.imgSrc} alt={`${product.title}상품`} />
-        <div className="title-wrapper">
-          <h3 className="title">{product.title}</h3>
+        <div className="top">
+          <ProductQuantity productId={productId} cartQuantity={cart ? cart.quantity : 0}>
+            <p className="title">{product.title}</p>
+          </ProductQuantity>
         </div>
         <div className="bottom flex-row-space-between">
           <p>가격</p>
